tried to refactor to new files, so i could merge to main. Some trouble resolving IDs
