+++ app/src/main/AndroidManifest.xml
@@ -23,6 +23,11 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".FirebaseDemoActivity"
+            android:exported="false"
+            android:theme="@style/Theme.MarioKartTournamentOrganizer">
+        </activity>
     </application>
 
 </manifest>
